reorganizing tiptap editor
