+++ backend/src/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -11,9 +11,7 @@
       "url": "https://khouryodyssey.org"
     },
     "license": null,
-
-    "x-generation-date": "2024-11-21T14:58:10.601Z"
-
+    "x-generation-date": "2024-12-23T14:10:27.344Z"
   },
   "x-strapi-config": {
     "plugins": [
