+++ frontend/components/draft/lesson/lesson-renderer.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { Lesson } from "@/types";
-import TipTap from "@/components/ui/tiptap/tiptap";
 import { ExpandableEditor } from "@/components/draft/lesson/blocks/expandable";
 import { VideoEditor } from "@/components/draft/lesson/blocks/video";
 import { CalloutEditor } from "./blocks/callout";
@@ -18,6 +17,7 @@ import { deleteLesson } from "@/lib/actions";
 import { Separator } from "@/components/ui/separator";
 import { toast } from "sonner";
 import { useTransition } from "react";
+import { LessonNameInput } from "@/components/ui/tiptap/lesson-name-input";
 
 export function LessonRenderer({
   lesson,
@@ -152,9 +152,8 @@ export function LessonRenderer({
   return (
     <>
       <div className="flex flex-col justify-center items-center border border-slate-200 rounded-md pt-4 px-4 pb-7 mb-5">
-        <TipTap
+        <LessonNameInput
           className="w-[700px] max-w-2xl mb-3"
-          variant="lesson-name"
           initialContent={"<h1>" + name + "</h1>"}
           updateContent={(content: string) => {
             content = htmlToText(content);
