+++ frontend/components/draft/metadata/droplet-name.tsx
@@ -4,7 +4,7 @@ import { Input } from "../../ui/input";
 import { useState } from "react";
 import { Label } from "../../ui/label";
 import { useDropletUpdate } from "./hooks/useDropletUpdate";
-import Tiptap from "../../ui/tiptap/tiptap";
+import { DropletNameInput } from "@/components/ui/tiptap/droplet-name-input";
 
 export function DropletName({
   startingName,
@@ -30,10 +30,9 @@ export function DropletName({
       <Label htmlFor="name" className="font-bold pb-4" hidden>
         Droplet Name
       </Label>
-      <Tiptap
+      <DropletNameInput
         updateContent={updateName}
         initialContent={`<h1>${name}</h1>`}
-        variant="droplet-name"
       />
 
       {error && <div className="text-red-500 mt-2">{error}</div>}
