+++ frontend/components/draft/metadata/overview.tsx
@@ -3,6 +3,8 @@
 import TipTap from "../../ui/tiptap/tiptap";
 import { debounce } from "lodash";
 import { useDropletUpdate } from "./hooks/useDropletUpdate";
+import { Droplet } from "lucide-react";
+import { DropletOverviewInput } from "@/components/ui/tiptap/droplet-overview-input";
 
 export function Overview({
   dropletId,
@@ -16,10 +18,9 @@ export function Overview({
   return (
     <section className="w-full max-w-2xl">
       <h2 className="text-2xl font-bold text-slate-900">Overview</h2>
-      <TipTap
-        updateContent={(content: any) => handleChange({ overview: content })}
+      <DropletOverviewInput
+        updateContent={(content: string) => handleChange({ overview: content })}
         initialContent={initialContent}
-        variant="droplet-overview"
       />
       {error && <div className="text-red-500 mt-2">{error}</div>}
     </section>
