+++ frontend/components/ui/tiptap/custom-image.tsx
