+++ frontend/components/ui/tiptap/toolbar/tools/ordered-list-tool.tsx
@@ -1,0 +1,17 @@
+import { cn } from "@/lib/utils";
+import { Editor } from "@tiptap/react";
+import { ListOrderedIcon } from "lucide-react";
+
+export default function OrderedListTool({ editor }: { editor: Editor }) {
+  return (
+    <button
+      onClick={() => editor.chain().focus().toggleOrderedList().run()}
+      className={cn(
+        editor?.isActive("orderedList") ? "bg-slate-200" : "",
+        "p-2.5 rounded-md border border-transparent hover:border-slate-200",
+      )}
+    >
+      <ListOrderedIcon size={17} />
+    </button>
+  );
+}
